move language registry
