+++ i18n/index.ts
@@ -3,7 +3,7 @@
  *
  * To add a new language:
  * 1. Add the language code to SUPPORTED_LANGUAGES in domain/language.ts
- * 2. Create a translation file (e.g., `shared/i18n/es.ts`) with `const es: Translations = { ... }`
+ * 2. Create a translation file (e.g., `i18n/es.ts`) with `const es: Translations = { ... }`
  * 3. Import and add the translation to the `translations` record below
  * 4. Add language metadata to LANGUAGE_OPTIONS in entrypoints/popup/views/settings/LanguageSection.tsx
  * 5. Run `npm run compile` - TypeScript will catch any missing keys
