+++ infrastructure/storage/__tests__/translations.test.ts
@@ -1,8 +1,8 @@
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
 import { fakeBrowser } from 'wxt/testing/fake-browser';
 import { storage } from 'wxt/utils/storage';
+import { translations } from '@/i18n';
 import { STORAGE_KEYS } from '@/infrastructure/storage/storage-keys';
-import { translations } from '@/shared/i18n';
 import { createDeferred } from '@/test/utils/deferred';
 import { getStoredTranslations } from '../translations';
