+++ infrastructure/storage/translations.ts
@@ -1,7 +1,7 @@
 import { storage } from '#imports';
 import { getSupportedLanguage, type Language } from '@/domain/language';
+import { type Translations, translations } from '@/i18n';
 import { detectBrowserLanguage } from '@/infrastructure/browser/language';
-import { type Translations, translations } from '@/shared/i18n';
 import { STORAGE_KEYS } from './storage-keys';
 
 export async function getStoredTranslations(): Promise<Translations> {
